requirements.txt
asgiref==3.8.1
Django==5.0.6
django-background-tasks-updated==1.2.8
django-compat==1.0.15
django-cors-headers==4.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
ruff==0.5.0
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
